# Conflicts:
#	core/src/main/java/com/graphhopper/coll/GHTreeMapComposed.java
#	reader-osm/src/test/java/com/graphhopper/reader/osm/GraphHopperOSMTest.java
#	tools/src/main/java/com/graphhopper/tools/Measurement.java

+++ reader-gtfs/src/main/java/com/graphhopper/gtfs/fare/TicketPurchaseScoreCalculator.java
@@ -29,11 +29,11 @@ class TicketPurchaseScoreCalculator {
         int nMoreTransfers = 0;
     }
 
-    int calculateScore(TicketPurchase ticketPurchase) {
-        int cost = 0;
+    double calculateScore(TicketPurchase ticketPurchase) {
+        double cost = 0;
         for (Ticket ticket : ticketPurchase.getTickets()) {
-            cost -= (int) ticket.getFare().fare_attribute.price;
+            cost -= ticket.getFare().fare_attribute.price;
         }
-        return cost - ticketPurchase.getNSchwarzfahrTrips() * 60;
+        return cost - ticketPurchase.getNSchwarzfahrTrips() * 60.0;
     }
 }
